Commonly used

More Internal

KM3Acoustics.ASignal — Type

ASignal is a custom type with four fields to store all the information inside the raw acoustic binary files.

dom_id::Int32 ID of the module

utc_seconds:: UInt32 storing the first 4 Bytes and is a UNIX time stamp

ns_cycles:: UInt32 storing the second 4 Bytes

samples:: UInt32 storing the third 4 Bytes, corresponding to the number of data points accuired during the measring window

pcm:: Vector of Float32 of length frame_length, storing all other 4 Byte blocks. Each entry is a data point of the acoustic signal.

KM3Acoustics.Detector — Type

A KM3NeT detector.

KM3Acoustics.Detector — Method

function Detector(filename::AbstractString)

Create a Detector instance from a DETX file.

KM3Acoustics.Detector — Method

function Detector(io::IO)

Create a Detector instance from an IO stream.

KM3Acoustics.DetectorModule — Type

Either a base module or an optical module. A non-zero status means the module is "not OK". Individual bits can be read out to identify the problem (see definitions/module_status.jl for the bit positions and check them using the nthbitset() function).

KM3Acoustics.Emitter — Type

The tripods in the seabed are Emitters of acoustics signals.

KM3Acoustics.Event — Type

An accoustic event is a collection, of a minimum number, of accoustic signals emmited from one tripod, gathered from multiple modules during a certain period of time.

KM3Acoustics.EventHeader — Type

EventHeader contains the information aboout detector, emitter, run.

KM3Acoustics.Hydrophone — Type

A hydrophone, typically installed in the base module of a KM3NeT detector's string.

KM3Acoustics.Location — Type

A module's location in the detector where string represents the detection unit identifier and floor counts from 0 from the bottom to top. Base modules are sitting on floor 0 and optical modules on floor 1 and higher.

KM3Acoustics.PMT — Type

The photomultiplier tube of an optical module. The id stands for the DAQ channel ID.

A non-zero status means the PMT is "not OK". Individual bits can be read out to identify the problem (see definitions/pmt_status.jl for the bit positions and check them using the nthbitset() function).

KM3Acoustics.Precalibration — Type

struct Precalibration <: Function

Data structure which stores all information needed for the precalibration process.

KM3Acoustics.Precalibration — Method

function Precalibration(detector_pos, hydrophones::OrderedDict{Int32, Hydrophone}, key_fixhydro::Int, events::Vector{Event}, emitters::Dict{Int8, Emitter})

Method to set up the precalibration data type.

KM3Acoustics.RawToashort — Type

RawToashort is an input data type which stores all information from the toashort file.

KM3Acoustics.Receiver — Type

Receivers are either DOMs with an piezo element or a baseunit with a hydrophone.

KM3Acoustics.StringCalibration — Type

Stores all the information needed for optimization procedure.

KM3Acoustics.StringCalibration — Method

Additional contructor for StringCalibration.

KM3Acoustics.Toashort — Type

Toashort is an input data type which stores all important information from the toashort file.

KM3Acoustics.ToyStringCalibration — Type

Stores all the information needed for toy optimization procedure.

KM3Acoustics.ToyStringCalibration — Method

Additional contructor for ToyStringCalibration.

KM3Acoustics.Transmission — Type

Datatype which has all information of one Transmission which is later needed for the fitting procedure.

KM3Acoustics.TriggerParameter — Type

Certain parameters which define an acoustic event.

KM3Acoustics.Tripod — Type

A tripod installed on the seabed which sends acoustic signals to modules.

KM3Acoustics.Waveform — Type

Waveform translates Emitter ID to Tripod ID.

Base.getindex — Method

Base.getindex(d::DetectorModule, i) = d.pmts[i+1]

The index in this context is the DAQ channel ID of the PMT, which is counting from 0.

Base.isless — Method

function isless(A::Transmission, B::Transmission)

Compares two transmissions. Necessary for sorting transmissions in the right way: Sort first by earliest TOA and if TOAs are equal sort first by higher Quality factor Q.

Base.read — Function

function read(filename::AbstractString,T::Type{ASignal}, overlap::Int=DAQ_ADF_ANALYSIS_WINDOW_OVERLAP)

Reads in a raw binary acoustics file.

Base.read — Method

function read(filename::AbstractString, Emitter, detector::Detector)

Reads in the tripod file, and return them as Emitters, where the positions are references in the detector coordinate system.

Base.read — Method

function read(filename::AbstractString, T::Type{Hydrophone})

Reads a vector of Hydrophones from an ASCII file.

Base.read — Method

function read(filename::AbstractString, T::Type{Toashort}, run::Int)

Reads a HDF5 file of toashorts.

Base.read — Method

function read(filename::AbstractString, T::Type{TriggerParameter})

Reads the 'acousticstriggerparameters.txt' file.

Base.read — Method

function read(filename:AbstractString, T::Type{Tripod})

Reads a vector of Tripods from an ASCII file.

Base.read — Method

function read(filename::AbstractString, T::Type{Waveform})

Reads the waveform ASCII file.

Base.read — Method

function read(file::HDF5.File, T::Type{Toashort}, run::Int)

Reads a HDF5 file of toashorts.

Base.write — Method

function write(filename::AbstractString, tripods::Dict{Int8, Tripod})

Writes the position of tripods out into an ASCII file.

Base.write — Method

function write(io::IO, d::Detector; version=:same)

Writes the detector to a DETX formatted file. The target version can be specified via the version keyword. Note that if converting to higher versions, missing parameters will be filled with reasonable default values. In case of downgrading, information will be lost.

KM3Acoustics._extract_comments — Method

function _extract_comments(lines<:Vector{AbstractString}, prefix<:AbstractString)

Returns only the lines which are comments, identified by the prefix. The prefix is omitted.

KM3Acoustics.acoustic_event — Method

function acoustic_event(detector, emitter, receivers, toe, run; p=0.8)

Simulates an acoustic event from one emitter to the receivers. The receiver gets the signal with a probability of p. Returns a RawToashort.

KM3Acoustics.calc_pos — Method

function calc_traveltime(dx, dy, l, a, b, basepos, emitterpos)

Calculates the position of a module located an some string at length l measured from the basemodule.

KM3Acoustics.calc_traveltime — Method

function calc_traveltime(dx, dy, l, a, b, basepos, emitterpos)

Calculates the traveltime of a signal of an emitter to a module located at the position length l at some string.

KM3Acoustics.check_basemodules — Method

function check_modules!(receivers, detector, hydrophones)

Checks if the modules in detector have hydrophones or piezos, if they have they will be written in receiver and emitters dicts.

KM3Acoustics.emitter_to_tripod — Method

function emitter_to_tripod(emitters::Dict{Int8, Emitter}, detector)

Takes in a dictionary of emitters and returns a vector of tripods.

KM3Acoustics.generate_startvalues — Method

function generate_startvalues(hydrophones, emitters, fixhydrophones, fixemitters, fitevents)

Generates starting values for the fitting procedure. The startvalues are essentially some old positions of hydrophones and tripods. The start values for time of emission we take the mean TOE of an acoustic event.

KM3Acoustics.get_basemodules — Method

function get_basemodules(modules, hydrophones)

Returns all basemodules and positions.

KM3Acoustics.get_hydrophones — Method

function get_hydrophones(filename::AbstractString, detector::Detector, events::Vector{Event})

Takes in a filename relating to hydrophones, a detector file and events from eventbuilder. Outputs an ordered dictionary of hydrophones which are involved in some event and in detector.

KM3Acoustics.get_opt_modules — Method

function get_opt_modules(p, pc)

Return two dictionaries of hydrophones and emitters, with precalibrated positions.

KM3Acoustics.group_events — Method

Groups events during a certain time window. Within this time window there should be at least two events from different tripods.

KM3Acoustics.group_fitevents — Method

function group_fitevents(events::Vector{Event}, nevents)

Groups a vector of sorted events, by the emitters which send the event. nevents is the maximum number of events we use for fitting per emitter.

KM3Acoustics.hydrophoneenabled — Method

function hydrophonenabled(m::DetectorModule)

Return true if the hydrophone is enabled, false otherwise.

KM3Acoustics.init_realdetector — Method

function init_realdetector(basemodules, modules)

Initializes an detector without calibration.

KM3Acoustics.init_toydetector — Method

function init_toydetector(basemodules, modules)

Initializes an detector without calibration.

KM3Acoustics.inverse_waveforms — Method

function inverse_waveforms(waveform::Waveform)

Return an inverse ( not injective ) map of the waveform.txt file.

KM3Acoustics.lookuptable_hydrophones — Method

function lookuptable_hydrophones(hydrophones::OrderedDict{Int32, Hydrophone}, key_fixhydro::Int32)

Returns a dictionary which maps the keys to the position in which they are sorted. First key -> 1, second key -> 2, ..

KM3Acoustics.natural — Method

function natural(x, y)

For sorting arrays of strings, including numbers in the strings. Usage: x = ["Foo101", "Foo13"] -> sort(x, lt=natural) -> ["Foo13", "Foo101"]

KM3Acoustics.nthbitset — Method

nthbitset(n, a) = !Bool((a >> (n - 1)) & 1)

Return true if the n-th bit of a is set, false otherwise.

KM3Acoustics.overlap — Method

function overlap(A::Event, B::Event, tmax::Float64)

Compares two events, which are already sorted by TOE, to check for overlap. If TOE of last signal of first event bigger than TOE of first signal of second event minus TMAX there is an overlap between signals.

KM3Acoustics.parse_runs — Method

function parse_runs(r)

Helper function to parse runs in eventbuilder script.

KM3Acoustics.piezoenabled — Method

function piezoenabled(m::DetectorModule)

Return true if the piezo is enabled, false otherwise.

KM3Acoustics.preprocess — Method

function preprocess(raw_signals)

Calculates UTC TOA and removes duplicate data.

KM3Acoustics.read_events — Method

function read_events(filename::AbstractString, run)

Reads in all events of a certain run from event.h5 file.

KM3Acoustics.read_events — Method

function read(file::HDF5.File, T::Type{Event}, run)

Reads in all events of a certain run from event.h5 file.

KM3Acoustics.read_toa — Method

function read_toa(filename::AbstractString, run::Int)

Acts as a function barrier. Opens the H5 File and reads the dataset for a specific group.

KM3Acoustics.rerotate_detector — Method

function rerotate_detector(hydrophones, emitters, ϕ)

Rotates the whole detector back to the originally position.

KM3Acoustics.rotate_detector — Method

function rotate_detector(hydrophones, emitters, pos)

Rotates the whole detector, including hydrophones and emitters, such that the position which is given as an argument lies in yz plane.

KM3Acoustics.save_events — Method

function save_events(events)

Output events as HDF5 file.

KM3Acoustics.select_fitevents — Method

function select_fitevents(events, mask)

Selects from the grouped events only certain events from different emitters.

KM3Acoustics.sort_fitevents — Method

function sort_events_qualityfactor(events::Vector{Event})

Sorts a vector of events. Events with highest mean qualityfactor come first.

KM3Acoustics.split_p — Method

function split_p(ps, hydrophones, emitters, numevents)

Splits the arguments of the fitting function into position of hydrophones, emitters and TOEs.

KM3Acoustics.string_length — Method

function string_length(dx, dy, z, a, b)

Given the height of a module located in some string, calculates the length of the string from the basemodule to the module.

KM3Acoustics.toy_calc_pos — Method

function toy_calc_pos(θ, ϕ, l)

Given the length of the string up to the module, returns the position of the module, relative to the basemodule

KM3Acoustics.toy_calc_traveltime — Method

function toy_calc_toa(θ, ϕ, l, basepos, emitterpos)

Given the length of the string up to the module, the angles of the strings, returns the traveltime between the given module and the emitter.

KM3Acoustics.traveltime — Method

traveltime(R::Float64, z1::Float64, z2::Float64, z_reference::Float64)

For a given distance between to points and their heights, returns the time for an acoustic signal to travel between the points. Heights are referenced to z_reference.

KM3Acoustics.traveltime — Method

traveltime(A, B, z_reference::Float64)

For a given module, tripod, emitter or receiver returns the time it takes for the signals to travel from the tripod to the module. The z positions are referenced to z_reference.

KM3Acoustics.traveltime — Method

function traveltime(A, B)

For a given module, tripod, emitter or receiver returns the time it takes for the signals to travel from the tripod to the module.

KM3Acoustics.traveltime — Method

function traveltime(x::Position, y::Position)

Given to Positions, returns the time for an acoustic signal to travel between positions.

KM3Acoustics.traveltime — Method

function traveltime(x::Position, y::Position, z_reference::Float64)

Given to Positions, returns the time for an acoustic signal to travel between positions. Heights are referenced to z_reference.

KM3Acoustics.traveltime — Method

traveltime(R::Float64, z1::Float64, z2::Float64)

For a given distance between to points and their heights, returns the time for an acoustic signal to travel between the points.

KM3Acoustics.tripod_to_emitter — Method

function tripod_to_emitter(tripods, detector)

Tripods position reference gets changed, such that the position is measured from the position of the detector. Returns a dictionary.

KM3Acoustics.unwrap — Method

function unwrap(p, mapping, hydrophones, emitters)

Helper function for mapping the arguments of the fitting function.

KM3Acoustics.velocity — Method

function get_velocity(T::DetectorModule; ref_soundvelocity=ref_soundvelocity, dv_dz=dv_dz)

For a given module or tripod returns the speed of sound at the depth of the module or tripod.

KM3Acoustics.velocity — Method

function get_velocity(z::T; ref_soundvelocity=ref_soundvelocity, dv_dz=dv_dz)

For a given depth z, returns the speed of sound at that depth.